Added validation and edit searching. Cant insert tabulation symbol for category menu=)

+++ src/main/java/com/azino/project/controller/CategoryController.java
@@ -2,11 +2,20 @@ package com.azino.project.controller;
 
 import com.azino.project.model.Category;
 import com.azino.project.service.CategoryService;
+import com.azino.project.util.WebUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ui.ModelMap;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import javax.servlet.http.HttpServletResponse;
+import javax.validation.Valid;
+import java.io.IOException;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("categories")
@@ -25,7 +34,12 @@ public class CategoryController {
 
     @PostMapping
     public ModelAndView add(@RequestParam(value = "name", required = true) String name,
-                            @RequestParam(value = "category", required = false) Category category) {
+                            @RequestParam(value = "category", required = false) Category category,
+                            RedirectAttributes redirectAttributes) {
+        if (name.length() < Category.NAME_MIN_LENGTH || name.length() > Category.NAME_MAX_LENGTH) {
+            redirectAttributes.addAttribute("error", "name length must be in range " + Category.NAME_MIN_LENGTH + " to " + Category.NAME_MAX_LENGTH);
+            return new ModelAndView("redirect:/menu/categories/addCategory");
+        }
         if (category != null) {
             categoryService.add(name, category);
         } else {
@@ -41,9 +55,13 @@ public class CategoryController {
 
     @DeleteMapping("{id}")
     public String delete(@PathVariable Long id) {
-        Category category = categoryService.findById(id).get();
-        categoryService.delete(id);
-        return "Ok";
+        Optional<Category> categoryOptional = categoryService.findById(id);
+        if (categoryOptional.isPresent()) {
+            categoryService.delete(id);
+            return "Ok";
+        } else {
+            return "No object";
+        }
     }
 
     @GetMapping("update")
@@ -53,10 +71,28 @@ public class CategoryController {
 
     @PutMapping("{id}")
     @Transactional
-    public String update(@RequestBody Category category, @PathVariable Long id) {
-        Category newCategory = categoryService.findById(id).get();
-        newCategory.setName(category.getName());
-        return "Updated!";
+    public String update(@Valid @RequestBody Category category, BindingResult bindingResult, @PathVariable Long id, HttpServletResponse response) {
+        try {
+            if(id < 0){
+                response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Id must be greater than zero!");
+                return "Error!";
+            }
+            if (bindingResult.hasErrors()) {
+                StringBuilder error = WebUtils.bindingResultErrorsToString(bindingResult);
+                response.sendError(HttpServletResponse.SC_BAD_REQUEST, error.toString());
+                return "Error!";
+            }
+            Optional<Category> newCategoryOptional = categoryService.findById(id);
+            if (newCategoryOptional.isPresent()) {
+                newCategoryOptional.get().setName(category.getName());
+                return "Updated!";
+            } else {
+                response.sendError(HttpServletResponse.SC_BAD_REQUEST, "No object with this id!");
+                return "Error!";
+            }
+        } catch (IOException ioe) {
+            ioe.printStackTrace();
+        }
+        return "Error!";
     }
-
 }
