+++ src/main/resources/static/js/categories/delete.js
@@ -1,16 +1,20 @@
 $(document).ready(function () {
     $('#deleteBtn').click(function (e) {
         var category = parseInt($('input[name=category]:checked').val());
-        $.ajax({
-            url: '/categories/' + category,
-            type: 'DELETE',
-            contentType: 'application/json',
-            success: function (data, textStatus, jqXHR) {
-                alert('Response: data = { ' + data +
-                    ' }\n textStatus = { ' + textStatus +
-                    ' }\n jqXHR.status = { ' + jqXHR.status + ' }');
-                window.location.href = "/account";
-            }
-        });
+        if(!Number.isNaN(category)) {
+            $.ajax({
+                url: '/categories/' + category,
+                type: 'DELETE',
+                contentType: 'application/json',
+                success: function (data, textStatus, jqXHR) {
+                    alert('Response: data = { ' + data +
+                        ' }\n textStatus = { ' + textStatus +
+                        ' }\n jqXHR.status = { ' + jqXHR.status + ' }');
+                    window.location.href = "/account";
+                }
+            });
+        }else{
+            alert("Choose category");
+        }
     });
 });
